+++ greenfield/eldenring/data.py
@@ -5557,4 +5557,4 @@ GESTURE_AWARD_FLAGS = {
     60864: (115, 1075750828, 'O Mother'),
 }
 
-_GEN_STAMP = {'inputs_hash': 'sha256:4b932035b1be2c4decfe6b797d3ee4ac508e6d6a07dc375eb00f9d3a1dfd8f23', 'module': 'data.py', 'body_sha256': 'sha256:aaf758d0d2de92f871bfc3e6e212c7c75a912bbddafce9595a2a26b55ef326b0'}
+_GEN_STAMP = {'inputs_hash': 'sha256:a9d609ae85eaff317d2cad16fc7108a58179a973de420739cc36a0852a748c05', 'module': 'data.py', 'body_sha256': 'sha256:aaf758d0d2de92f871bfc3e6e212c7c75a912bbddafce9595a2a26b55ef326b0'}